don't remove applications if pipeline fails

+++ tests/e2e-demos/multi-component.go
@@ -52,6 +52,8 @@ var _ = framework.E2ESuiteDescribe(Label("e2e-demo"), func() {
 	var compNameGo string
 	var compNameNode string
 
+	var removeApplication = true
+
 	BeforeAll(func() {
 		// Check to see if the github token was provided
 		Expect(utils.CheckIfEnvironmentExists(constants.GITHUB_TOKEN_ENV)).Should(BeTrue(), "%s environment variable is not set", constants.GITHUB_TOKEN_ENV)
@@ -70,8 +72,10 @@ var _ = framework.E2ESuiteDescribe(Label("e2e-demo"), func() {
 
 	// Remove all resources created by the tests
 	AfterAll(func() {
-		Expect(fw.HasController.DeleteAllComponentsInASpecificNamespace(AppStudioE2EApplicationsNamespace, 30*time.Second)).To(Succeed())
-		Expect(fw.HasController.DeleteAllApplicationsInASpecificNamespace(AppStudioE2EApplicationsNamespace, 30*time.Second)).To(Succeed())
+		if removeApplication {
+			Expect(fw.HasController.DeleteAllComponentsInASpecificNamespace(AppStudioE2EApplicationsNamespace, 30*time.Second)).To(Succeed())
+			Expect(fw.HasController.DeleteAllApplicationsInASpecificNamespace(AppStudioE2EApplicationsNamespace, 30*time.Second)).To(Succeed())
+		}
 	})
 
 	It("Create Red Hat AppStudio Application", func() {
@@ -154,8 +158,17 @@ var _ = framework.E2ESuiteDescribe(Label("e2e-demo"), func() {
 	// Start to watch the pipeline until is finished
 	It("Wait for all pipelines to be finished", func() {
 
-		Expect(fw.HasController.WaitForComponentPipelineToBeFinished(compNameGo, testSpecification.Tests[0].ApplicationName, AppStudioE2EApplicationsNamespace)).To(Succeed(), "Failed component pipeline %v", err)
-		Expect(fw.HasController.WaitForComponentPipelineToBeFinished(compNameNode, testSpecification.Tests[0].ApplicationName, AppStudioE2EApplicationsNamespace)).To(Succeed(), "Failed component pipeline %v", err)
+		err := fw.HasController.WaitForComponentPipelineToBeFinished(compNameGo, testSpecification.Tests[0].ApplicationName, AppStudioE2EApplicationsNamespace)
+		if err != nil {
+			removeApplication = false
+		}
+		Expect(err).NotTo(HaveOccurred(), "Failed component pipeline %v", err)
+
+		err = fw.HasController.WaitForComponentPipelineToBeFinished(compNameNode, testSpecification.Tests[0].ApplicationName, AppStudioE2EApplicationsNamespace)
+		if err != nil {
+			removeApplication = false
+		}
+		Expect(err).NotTo(HaveOccurred(), "Failed component pipeline %v", err)
 
 	})
